// frontend/src/app/api/model/miniRouter.ts
// File generated using spec-to-code.sh, DO NOT EDIT MANUALLY.

/**
 * ADH6 API
 *
 * Contact: equipe@minet.net
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { MiniRouterLoan } from './miniRouterLoan';


export interface MiniRouter { 
    /**
     * The unique identifier of this mini-router
     */
    readonly id?: number;
    /**
     * The factory MAC address of the mini-router
     */
    hardwareMac: string;
    /**
     * The MAC address assigned by MiNET
     */
    mac?: string | null;
    /**
     * The IPv4 address of the mini-router
     */
    ip?: string | null;
    /**
     * The hardware model
     */
    model: MiniRouter.ModelEnum;
    /**
     * Whether the MiNET configuration is applied
     */
    configState: MiniRouter.ConfigStateEnum;
    /**
     * Free comment
     */
    comment?: string | null;
    /**
     * The room of the member of the current loan
     */
    readonly roomId?: number | null;
    /**
     * The room number of the member of the current loan
     */
    readonly roomNumber?: number | null;
    /**
     * Whether a returned loan still has its deposit held
     */
    readonly depositToRefund?: boolean;
    /**
     * The loan in progress, if any
     */
    readonly currentLoan?: MiniRouterLoan | null;
}
export namespace MiniRouter {
    export type ModelEnum = 'beryl_giga' | 'ar750_fast';
    export const ModelEnum = {
        BerylGiga: 'beryl_giga' as ModelEnum,
        Ar750Fast: 'ar750_fast' as ModelEnum
    };
    export type ConfigStateEnum = 'pimped' | 'not_pimped';
    export const ConfigStateEnum = {
        Pimped: 'pimped' as ConfigStateEnum,
        NotPimped: 'not_pimped' as ConfigStateEnum
    };
}